+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# REPLACE_GLOBAL_WITH_NEXT_VERSION
+# 11.0.0
 - MIG-951 - Improve the language migration from Shopware 6 to Shopware 6 so the default language is no longer overwritten
 - MIG-951 - [BREAKING] Changes the access modifier of the function `\SwagMigrationAssistant\Profile\Shopware6\Converter\LanguageConverter::convertData` from public to protected
 - MIG-943 - Fix migration of calculated shipping costs from Shopware 5